Fix issue/bug with button/links to Athletics YT channel...
When I added the new buttons linking to the Athletics YouTube channel,
I didn't catch that they where unclickable--obscured by an invisible element over top them.

Fixed by removing the `position: absolute;` used on the element holding the schedule-key and button.
(The absolute positioning was an aesthetic fix to lessen the gap between the table and the key.)